Migrate data-service to typescript
